+++ go/internal/site/site.go
@@ -105,6 +105,10 @@ func Install(logf execx.Logf, o Options) error {
 	if err := moodle.WriteConfig(o.Wwwroot); err != nil {
 		return err
 	}
+	logf("Writing the single-use login handler (mdldemo-login.php)")
+	if err := moodle.WriteSSOScript(); err != nil {
+		return err
+	}
 
 	docroot := moodle.Docroot()
 	logf("Configuring Apache (docroot " + docroot + ")")